% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rzeroconf.R
\docType{package}
\name{Rzeroconf-package}
\alias{Rzeroconf}
\alias{Rzeroconf-package}
\title{libgeos}
\description{
Based on the servus library from https://github.com/HBPVIS/Servus
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.github.com/rundel/Rzeroconf}
  \item Report bugs at \url{http://www.github.com/rundel/Rzeroconf/issues}
}

}
